Avoid calling host_init for USB peripherals
When `usb start` is called, the dwc2 driver will try to start every USB device as host first, even if it is explicitly configured as peripheral in the device tree (`dr_mode = "peripheral"`).

So to avoid am unwanted 15 seconds delay when initializing the usb (one second per channel = 1s x 15), this patch adds a check to the initialization, and will skip host initialization of the device is explicitly set as peripheral. The checking is already and correctly done similarly in the `drivers/usb/gadget/dwc2_udc_otg.c` driver.